requirements.txt
aiogram>=3.0.0
pymongo>=4.0.0
python-dotenv==1.1.1 
psutil==6.0.0